Add new variable to install gie-proxy
The new variable is gie_proxy_install (defaults to true). Through this
variable we can decide whether to install (clone the gx-it-proxy repo
and compile the Node JS modules) or not. We only want this to be installed on
the maintenance node without the systemd service and then on the
headnode we would need only the systemd service. So with this new variable we
can switch them ON and OFF depending on where we are installing. This
variable will not affect the current behaviour (default is set to true).
For other tasks in there is already a variable which we can use to
switch things ON and OFF.